refactoring xml based functionality into xmlfileprocessor service and creatign dedicated recordtypeservice
